pyproject.toml
[project]
name = "fulcrum"
version = "0.1.0"
description = "Executive GCP catalog CLI with rich TUI"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
  "typer",
  "textual",
  "rich",
  "pydantic>=2",
  "structlog",
  "jinja2",
  "pyyaml",
  "tomlkit",
  "pytest>=8.3.5",
  "pytest-cov>=5.0.0",
  "google-api-python-client>=2.141.0",
  "google-auth>=2.27.0",
  "google-auth-httplib2>=0.2.0",
  "google-cloud-storage>=2.18.0",
  "tenacity>=9.0.0",
  "responses>=0.25.0",
  "pytest-mock>=3.12.0",
  "requests>=2.32.0",
  "toml>=0.10.2",
]

[project.scripts]
fulcrum = "src.cli:main"

[tool.pytest.ini_options]
addopts = "-q"
testpaths = ["tests"]

[tool.mypy]
python_version = "3.8"
warn_redundant_casts = true
warn_unused_ignores = true
strict = true

[build-system]
requires = ["setuptools", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = ["."]
include = ["src*"]
